EDU: Adjusted FrameUpdate new paramaters
Debug: Started Joint Debug visuals

+++ src/ea/edu/EduScene.java
@@ -458,7 +458,7 @@ public class EduScene extends Scene implements KeyListener, MouseClickListener, 
 
         @Override
         public Class<?>[] getInvocationMethodParameters() {
-            return new Class<?>[] {Integer.class};
+            return new Class<?>[] {float.class};
         }
     }
